+++ lib/localize.js
@@ -5,24 +5,6 @@ import IntlMessageFormat from 'intl-messageformat';
 
 export const allowedTags = Object.freeze(['d2l-link', 'd2l-tooltip-help', 'p', 'br', 'b', 'strong', 'i', 'em', 'button']);
 
-const characterMap = new Map([
-	['\'', 'apostrophe'],
-	['&', 'ampersand'],
-	['*', 'asterisk'],
-	['\\', 'backslash'],
-	[':', 'colon'],
-	[',', 'comma'],
-	['>', 'greaterThan'],
-	['<', 'lessThan'],
-	['#', 'numberSign'],
-	['%', 'percentSign'],
-	['|', 'pipe'],
-	['?', 'questionMark'],
-	['"', 'quotationMark']
-]);
-const commonResources = new Map();
-export let commonResourcesImportCount = 0;
-
 const getDisallowedTagsRegex = allowedTags => {
 	const validTerminators = '([>\\s/]|$)';
 	const allowedAfterTriangleBracket = `/?(${allowedTags.join('|')})?${validTerminators}`;
@@ -86,17 +68,6 @@ export const getLocalizeClass = (superclass = class {}) => class LocalizeClass e
 		return formattedMessage;
 	}
 
-	localizeCharacter(char) {
-		if (!characterMap.has(char)) {
-			throw new Error(`localizeCharacter() does not support character: "${char}"`);
-		}
-		const value = this.localize(`intl-common:characters:${characterMap.get(char)}`);
-		if (value.length === 0) {
-			throw new Error('localizeCharacter() cannot be used unless loadCommon in localizeConfig is enabled');
-		}
-		return value;
-	}
-
 	localizeHTML(name, replacements = {}) {
 
 		const { language, value } = this.localize.resources?.[name] ?? {};
@@ -156,20 +127,9 @@ export const getLocalizeClass = (superclass = class {}) => class LocalizeClass e
 		}
 		if (Object.prototype.hasOwnProperty.call(this, 'getLocalizeResources') || Object.prototype.hasOwnProperty.call(this, 'resources')) {
 			const possibleLanguages = this._generatePossibleLanguages(config);
-			if (config?.importFunc) {
+			if (config.importFunc) {
 				const resourcesPromise = this.getLocalizeResources(possibleLanguages, config);
 				resourcesLoadedPromises.push(resourcesPromise);
-			}
-			if (config?.loadCommon) {
-				resourcesLoadedPromises.push(this.getLocalizeResources(possibleLanguages, {
-					importFunc: async(lang) => {
-						if (commonResources.has(lang)) return commonResources.get(lang);
-						const resources = (await import(`../lang/${lang}.js`)).default;
-						commonResourcesImportCount++;
-						commonResources.set(lang, resources);
-						return resources;
-					}
-				}));
 			}
 		}
 		return Promise.all(resourcesLoadedPromises);
